pyproject.toml
[project]
name = "fitness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.1",
    "crewai>=0.121.1",
    "fastapi>=0.115.9",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "openai>=1.75.0",
    "paper-qa>=5.21.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.11.5",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.23",
    "torch>=2.7.0",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.34.2",
]
